requirements.txt
annotated-types==0.7.0
anyio
argon2-cffi
argon2-cffi-bindings
arrow==1.3.0
asttokens
async-lru
attrs
Babel
beautifulsoup4
bleach
Brotli
certifi
cffi
charset-normalizer
click==8.1.7
comm
debugpy
decorator
defusedxml
executing
fastapi==0.115.4
fastjsonschema
filelock==3.16.1
fqdn==1.5.1
fsspec==2024.10.0
h11
httpcore
httpx
huggingface-hub==0.26.2
idna
ipykernel
ipython
ipywidgets==8.1.5
isoduration==20.11.0
jedi
Jinja2
json5
jsonpointer==3.0.0
jsonschema
jsonschema-specifications
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab_server
jupyterlab_widgets==3.0.13
markdown2==2.5.1
MarkupSafe
matplotlib-inline
mistune
mpmath==1.3.0
nbclient
nbconvert
nbformat
nest-asyncio
networkx==3.4.2
notebook
notebook_shim
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.3.3
opencv-python==4.10.0.84
overrides
packaging
pandocfilters
parso
pexpect
pillow==11.0.0
platformdirs
prometheus-client
prompt-toolkit
psutil
ptyprocess
pure-eval
pycparser
pydantic==2.9.2
pydantic_core==2.23.4
Pygments
PySocks
python-dateutil
python-json-logger
python-multipart==0.0.16
pytz
PyYAML
pyzmq
referencing
regex==2024.9.11
requests
rfc3339-validator
rfc3986-validator
rpds-py
safetensors==0.4.5
Send2Trash
six
sniffio
soupsieve
stack-data
starlette==0.41.2
sympy==1.13.1
terminado
timm==1.0.11
tinycss2
tokenizers==0.20.1
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tornado
tqdm==4.66.6
traitlets
transformers==4.46.1
triton==3.1.0
types-python-dateutil==2.9.0.20241003
typing_extensions
uri-template==1.3.0
urllib3
uvicorn==0.32.0
wcwidth
webcolors==24.8.0
webencodings==0.5.1
websocket-client
widgetsnbextension==4.0.13
